usersController,postsControllerへ名称変更 post_categories=>postCategoriesへ変更 postsControllerコード修正 => status(404)追加 statusのパラメータ[0,100,200]定数定義 トランザクション実装(create,updateメソッド) show,indexメソッドに名称変更
