+++ controllers/usersController.ts
@@ -1,18 +1,18 @@
 import { Request, Response } from "express";
-import posts from "../models/post";
-import post_categories from "../models/postCategory";
+import posts, { statusValues } from "../models/post";
+import postCategories from "../models/postCategory";
 import categories from "../models/category";
 
 export default {
   //authuserを返す
-  async findOne(req: any, res: Response) {
+  async show(req: any, res: Response) {
     const { authorize_token, ...user } = req.user.toJSON();
     res.status(200).json({ user });
   },
   //authuserに紐づくpostを返す
   async findPost(req: any, res: Response) {
     const { ...user } = req.user.toJSON();
-    const queryStatus = req.query.status ? req.query.status : [0, 100, 200];
+    const queryStatus = req.query.status ? req.query.status : statusValues;
     const id = user.id;
     const post = await posts.findAll({
       where: {
@@ -22,8 +22,8 @@ export default {
       raw: false,
       include: [
         {
-          model: post_categories,
-          as: "post_categories",
+          model: postCategories,
+          as: "postCategories",
           required: false,
           include: [
             {
