+++ models/postCategory.ts
@@ -2,12 +2,7 @@ import { Model, DataTypes }from "sequelize";
 import  sequelize  from '../middlewares/sequelize';
 
 export default class PostCategory extends Model {
-    public id!: number;
-    public postId!: number;
     public categoryId!: number;
-
-    public readonly createdAt!: Date;
-    public readonly updatedAt!: Date;
   }
 
 PostCategory.init(
@@ -26,7 +21,7 @@ PostCategory.init(
     },
   },
   {
-    tableName: "post_categories",
+    tableName: "postCategories",
     sequelize: sequelize,
   }
 );
