+++ routes/users.ts
@@ -1,9 +1,9 @@
 import express from "express";
-import usersController from "../controllers/users.controller";
+import usersController from "../controllers/usersController";
 
 const usersRouter = express.Router();
 
-usersRouter.get("/",usersController.findOne);
+usersRouter.get("/",usersController.show);
 usersRouter.get("/posts", usersController.findPost);
 
 export { usersRouter };
